backend/requirements.txt
Flask==2.2.3
werkzeug==2.2.2
mysql-connector-python==8.0.33
PyJWT==2.6.0   
python-dotenv==0.21.0